awesome: Replace runonce lua script by systemd-run
systemd-run can be used similar by creating a transient scope unit.
Scope units are not forked of by systemd-run and the environment is kept
(which we need for $DISPLAY..).
Also, by giving an explicit name of the unit, we ensure, it will not be
start twice, since the unit already exists.

This works just perfectly! and is much faster when booting and launching
big stuff like nm-applet and lxpolkit :)