quvi/xvidstage: Implement proper minjs unpacking
Dont try to be to clever and fetch the ip path data from the minified
js. Instead do a proper unpack and parse the url from the result.

Background:
The minified js might look slighty different if there are i.e.
redundacies in the ip (two times the same numerical component).